requirements.txt
flask==3.0.0
twilio==8.11.0
python-dotenv==1.0.0
gitpython==3.1.40
anthropic==0.34.0
